move boundaries into `boundaries.rs` and fix `ArithmeticSimdOperation` for i32, u32 and f32
